Issue 14 : jpn error due to https
Japan url changed from http to https and this caused the error.

I managed to reproduce the error by hand with an incomplete url (missing
file parameters). However the code was only working with interactivity
and not in the function.

The only way to fix this is to modify the url itself.

This was a great way to start using http requests with GET and POST